FinalProject/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# C:\GIT\Exercises\FinalProject\data.py: 6
# C:\GIT\Exercises\FinalProject\inference.py: 4
Pillow == 8.2.0

# C:\GIT\Exercises\FinalProject\data.py: 2
h5py == 3.1.0

# C:\GIT\Exercises\FinalProject\image_preprocessing.py: 3
# C:\GIT\Exercises\FinalProject\inference.py: 2,3
matplotlib == 3.3.4

# C:\GIT\Exercises\FinalProject\analysis.py: 3
# C:\GIT\Exercises\FinalProject\data.py: 7
# C:\GIT\Exercises\FinalProject\image_preprocessing.py: 2
# C:\GIT\Exercises\FinalProject\inference.py: 1
# C:\GIT\Exercises\FinalProject\model.py: 3
numpy == 1.19.5

# C:\GIT\Exercises\FinalProject\data.py: 3
opencv_python == 4.5.2.54

# C:\GIT\Exercises\FinalProject\analysis.py: 1
pandas == 1.2.4

# C:\GIT\Exercises\FinalProject\image_preprocessing.py: 1,4
scipy == 1.6.3

# C:\GIT\Exercises\FinalProject\analysis.py: 2
sklearn == 0.0

# C:\GIT\Exercises\FinalProject\data.py: 5
# C:\GIT\Exercises\FinalProject\model.py: 9,10
tensorflow_gpu == 2.5.0

# C:\GIT\Exercises\FinalProject\image_preprocessing.py: 5
tqdm == 4.61.1
